requirements.txt
APScheduler==3.6.3
autopep8==1.6.0
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.10
distro==1.6.0
et-xmlfile==1.1.0
google-api-core==2.4.0
google-api-python-client==2.36.0
google-auth==2.4.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
greenlet==1.1.2
gspread==5.1.1
httplib2==0.20.2
idna==3.3
numpy==1.22.0
oauth2client==4.1.3
oauthlib==3.1.1
openpyxl==3.0.9
pandas==1.3.5
protobuf==3.19.3
psycopg2-binary==2.9.3
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-telegram-bot==13.10
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.31
tabula-py==2.3.0
toml==0.10.2
tornado==6.1
tzdata==2021.5
tzlocal==4.1
uritemplate==4.1.1
urllib3==1.26.8
